Admin Authorization

+++ pages/api/products/[id].js
@@ -2,9 +2,11 @@ import dbConnect from '../../../util/mongo'
 import Product from '../../../models/Product'
 
 export default async function handler(req, res) {
-    const { method, query:{id} } = req
+    const { method, query:{id}, cookies } = req
 
-    dbConnect()
+    const token = cookies.token
+
+    await dbConnect()
 
     if(method === "GET") {
         try {
@@ -13,26 +15,14 @@ export default async function handler(req, res) {
         } catch (err) {
             res.status(500).json(err)
             console.error(err);
-            if (err.response) {
-                // The client was given an error response (5xx, 4xx)
-                console.error(err.response);
-                console.log(err.response.data);
-                console.log(err.response.status);
-                console.log(err.response.headers);
-            } else if (err.request) {
-                console.error(err.request.data);
-                console.error(err.request);
-                console.error(err.request.message);
-                // The client never received a response, and the request was never left
-            } else {
-                // Anything else
-            }   
         }
-
     }
-    if(method === "POST") {
+    if(method === "PUT") {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not Authorized!")
+        }
         try {
-            const product = await Product.create(req.body)
+            const product = await Product.findByIdAndUpdate(id, req.body, {new: true,})
             res.status(201).json(product)
         } catch (err) {
             res.status(500).json(err)
@@ -40,6 +30,9 @@ export default async function handler(req, res) {
         }
     }
     if(method === "DELETE") {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not Authorized!")
+        }
         try {
             await Product.findByIdAndDelete(id)
             res.status(201).json("The product has been deleted!")
